Rename the nms package to spigotnms for Spigot mapped NMS
